Reorganized iterator tests
